Moved the save form logic to a Redux action, using redux-thunk
