+++ client/main.js
@@ -2,10 +2,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux'
+import thunk from 'redux-thunk';
 import configureStore from 'state';
 import '../assets/stylesheets/style.scss';
 import initializeState from './lib/initialize-state';
-import saveForm from './lib/save-form';
 import './lib/calypso-boot';
 import { translate as __ } from 'lib/mixins/i18n';
 
@@ -18,9 +18,7 @@ const {
 	nonce,
 } = wcConnectData;
 
-const store = configureStore( initializeState( formSchema, formData ) );
-
-const saveFormData = ( setIsSaving, setSuccess, setError, filterStoreOnSave ) => saveForm( setIsSaving, setSuccess, setError, callbackURL, nonce, filterStoreOnSave( store ) );
+const store = configureStore( initializeState( formSchema, formData ), thunk.withExtraArgument( { callbackURL, nonce } ) );
 
 window.addEventListener( 'beforeunload', ( event ) => {
 	if ( store.getState().form.pristine ) {
@@ -41,7 +39,6 @@ let render = () => {
 				storeOptions={ storeOptions }
 				schema={ formSchema }
 				layout={ formLayout }
-				saveFormData={ saveFormData }
 			/>
 		</Provider>,
 		rootEl
