+++ client/state/settings/actions.js
@@ -1,3 +1,10 @@
+import saveForm from 'lib/save-form';
+import isString from 'lodash/isString';
+import isArray from 'lodash/isArray';
+import { translate as __ } from 'lib/mixins/i18n';
+import * as FormActions from 'state/form/actions';
+import * as NoticeActions from 'state/notices/actions';
+
 export const UPDATE_SETTINGS_FIELD = 'UPDATE_SETTINGS_FIELD';
 export const ADD_SETTINGS_OBJECT_FIELD = 'ADD_SETTINGS_OBJECT_FIELD';
 export const REMOVE_SETTINGS_OBJECT_FIELD = 'REMOVE_SETTINGS_OBJECT_FIELD';
@@ -59,3 +66,35 @@ export const updateSettingsArrayFieldItem = ( settings_key, index, item ) => ( {
 	index,
 	item,
 } );
+
+export const submit = ( silent ) => ( dispatch, getState, { callbackURL, nonce } ) => {
+	silent = ( true === silent );
+	const setIsSaving = ( value ) => dispatch( FormActions.setFormProperty( 'isSaving', value ) );
+	const setSuccess = ( value ) => {
+		dispatch( FormActions.setFormProperty( 'success', value ) );
+		if ( ! silent && true === value ) {
+			dispatch( NoticeActions.successNotice( __( 'Your changes have been saved.' ), {
+				duration: 2250,
+			} ) );
+		}
+	};
+	const setError = ( value ) => {
+		dispatch( FormActions.setFormProperty( 'errors', value ) );
+
+		if ( ! silent ) {
+			if ( isString( value ) ) {
+				NoticeActions.errorNotice( value, {
+					duration: 7000,
+				} );
+			}
+
+			if ( isArray( value ) ) {
+				NoticeActions.errorNotice( __( 'There was a problem with one or more entries. Please fix the errors below and try saving again.' ), {
+					duration: 7000,
+				} );
+			}
+		}
+	};
+
+	saveForm( setIsSaving, setSuccess, setError, callbackURL, nonce, getState().settings );
+};
