+++ client/views/shipping/index.js
@@ -1,13 +1,12 @@
 import React, { PropTypes } from 'react';
 import WCCSettingsForm from 'components/wcc-settings-form';
 
-const Settings = ( { storeOptions, schema, layout, saveFormData } ) => {
+const Settings = ( { storeOptions, schema, layout } ) => {
 	return (
 		<WCCSettingsForm
 			storeOptions={ storeOptions }
 			schema={ schema }
 			layout={ layout }
-			saveFormData={ saveFormData }
 		/>
 	);
 };
@@ -16,7 +15,6 @@ Settings.propTypes = {
 	storeOptions: PropTypes.object.isRequired,
 	schema: PropTypes.object.isRequired,
 	layout: PropTypes.array.isRequired,
-	saveFormData: PropTypes.func.isRequired,
 };
 
 export default Settings;
